Ignore ID3 Timestamp data if we've received segments without it
Remove ID3 dependencies when playing back audio streams.  The iHeartRadio stream occasionally had segments which included timeStamp data in its ID3.  This caused the stream to write audio as the wrong time and appear to stall.  Instead, we will simply write the audio in at the timeOffset that is provided into the stream.

JW7-3452